postgres -> new import style
