+++ splink/duckdb/duckdb_linker.py
@@ -4,7 +4,7 @@ from ..exceptions import SplinkDeprecated
 from .linker import DuckDBLinker, DuckDBLinkerDataFrame  # noqa: F401
 
 warnings.warn(
-    "Importing directly from `splink.duckdb.linker` "
+    "Importing directly from `splink.duckdb.duckdb_linker` "
     "is deprecated and will be removed in Splink v4. "
     "Please import from `splink.duckdb.linker` going forward.",
     SplinkDeprecated,
