+++ splink/postgres/postgres_comparison_level_library.py
@@ -1,13 +1,12 @@
-from ..comparison_level_composition import and_, not_, or_  # noqa: F401
-from .postgres_helpers.postgres_comparison_imports import (  # noqa: F401
-    array_intersect_level,
-    columns_reversed_level,
-    datediff_level,
-    distance_function_level,
-    distance_in_km_level,
-    else_level,
-    exact_match_level,
-    levenshtein_level,
-    null_level,
-    percentage_difference_level,
+import warnings
+
+from ..exceptions import SplinkDeprecated
+from .comparison_level_library import *  # noqa: F403
+
+warnings.warn(
+    "Importing directly from `splink.postgres.postgres_comparison_level_library` "
+    "is deprecated and will be removed in Splink v4. "
+    "Please import from `splink.postgres.comparison_level_library` going forward.",
+    SplinkDeprecated,
+    stacklevel=2,
 )
