+++ splink/spark/spark_linker.py
@@ -4,7 +4,7 @@ from ..exceptions import SplinkDeprecated
 from .linker import SparkDataframe, SparkLinker  # noqa: F401
 
 warnings.warn(
-    "Importing directly from `splink.spark.linker` "
+    "Importing directly from `splink.spark.spark_linker` "
     "is deprecated and will be removed in Splink v4. "
     "Please import from `splink.spark.linker` going forward.",
     SplinkDeprecated,
