+++ splink/sqlite/sqlite_comparison_library.py
@@ -4,7 +4,7 @@ from ..exceptions import SplinkDeprecated
 from .comparison_library import *  # noqa: F403
 
 warnings.warn(
-    "Importing directly from `splink.sqlite.comparison_library` "
+    "Importing directly from `splink.sqlite.sqlite_comparison_library` "
     "is deprecated and will be removed in Splink v4. "
     "Please import from `splink.sqlite.comparison_library` going forward.",
     SplinkDeprecated,
